+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 ### Add
 - It is possible to create a sp using a keras models and nods_info file to format the distribution. 
 - Local distributor that connects to a NOA in the cloud in order to register the synaptic process(sp) and model.
+- Fixed neuron distributor error when sending neurons to more than 4 nods